seperate logic into two controller files
