+++ ToDoList/Controllers/HomeControllers.cs
@@ -1,6 +1,4 @@
 using Microsoft.AspNetCore.Mvc;
-using ToDoList.Models;
-using System.Collections.Generic;
 
 namespace ToDoList.Controllers
 {
@@ -9,19 +7,7 @@ namespace ToDoList.Controllers
     [HttpGet("/")]
     public ActionResult Index()
     {
-      List<Item> allItems = allItems.GetAll();
-      return View(allItems);
-    }
-    [HttpGet("/items/new")]
-    public ActionResult CreateForm()
-    {
       return View();
-    }
-    [HttpPost("/items")]
-    public ActionResult Create(string description)
-    {
-      Item myItem = new Item(description);
-      return RedirectToAction("Index");
     }
   }
 }
