+++ ToDoList/Controllers/ItemsController.cs
@@ -1,0 +1,27 @@
+using Microsoft.AspNetCore.Mvc;
+using ToDoList.Models;
+using System.Collections.Generic;
+
+namespace ToDoList.Controllers
+{
+  public class ItemsController : Controller
+  {
+    [HttpGet("/items")]
+    public ActionResult Index()
+    {
+      List<Item> allItems = Items.GetAll();
+      return View(allItems);
+    }
+    [HttpGet("/items/new")]
+    public ActionResult CreateForm()
+    {
+      return View();
+    }
+    [HttpPost("/items")]
+    public ActionResult Create(string description)
+    {
+      Item myItem = new Item(description);
+      return RedirectToAction("Index");
+    }
+  }
+}
