+++ ToDoList/Views/Home/Index.cshtml
@@ -1,14 +1,3 @@
-@using ToDoList.Models;
-
-<h1>To Do List</h1>
-  @if (Model.Count == 0)
-  {
-    <p>There are no items in the list.</p>
-  }
-<ul>
-  @foreach (Item item in Model)
-  {
-      <li>@item.Description</li>
-  }
-</ul>
-<a href="/items/new">Add a new item.</a>
+<h1>Welcome to the To Do List!</h1>
+<a href="/items/new">Add a new item</a>
+<a href="/items">See list of items</a>
