+++ ToDoList/Views/Items/Index.cshtml
@@ -1,0 +1,14 @@
+@using ToDoList.Models;
+
+<h1>To Do List</h1>
+  @if (Model.Count == 0)
+  {
+    <p>There are no items in the list.</p>
+  }
+<ul>
+  @foreach (Item item in Model)
+  {
+      <li>@item.Description</li>
+  }
+</ul>
+<a href="/items/new">Add a new item.</a>
